refactor: 예외 메시지 클래스 명 수정 및 패키지 이동
- CommonErrorMessage에서 SharedErrorMessage로 변경
- 구체 클래스를 도메인 하위로 이동
