+++ src/main/java/lotto/domain/shared/LottoNumber.java
@@ -1,7 +1,5 @@
 package lotto.domain.shared;
 
-import lotto.common.CommonErrorMessage;
-
 public record LottoNumber(int number) {
     private static final int MINIMUM_NUMBER = 1;
     private static final int MAXIMUM_NUMBER = 45;
@@ -12,7 +10,7 @@ public record LottoNumber(int number) {
 
     private void validateNumberInRange(final int number) {
         if (number < MINIMUM_NUMBER || number > MAXIMUM_NUMBER) {
-            throw new IllegalArgumentException(CommonErrorMessage.LOTTO_NUMBER_OUT_OF_RANGE.message()
+            throw new IllegalArgumentException(SharedErrorMessage.LOTTO_NUMBER_OUT_OF_RANGE.message()
                     .formatted(MINIMUM_NUMBER, MAXIMUM_NUMBER));
         }
     }
