+++ src/main/java/lotto/domain/shared/Money.java
@@ -1,7 +1,5 @@
 package lotto.domain.shared;
 
-import lotto.common.CommonErrorMessage;
-
 public record Money(int amount) {
     private static final int LOWER_LIMIT = 0;
 
@@ -11,7 +9,7 @@ public record Money(int amount) {
 
     private void validateAmount(int amount) {
         if (amount < LOWER_LIMIT) {
-            throw new IllegalArgumentException(CommonErrorMessage.MONEY_NEGATIVE_VALUE.message());
+            throw new IllegalArgumentException(SharedErrorMessage.MONEY_NEGATIVE_VALUE.message());
         }
     }
 }
